Don't suggest token reuse for double-puppeting
While this mostly works, it means that there are now two clients A and B acting
as the same logical [device]. The two clients are indistinguishable to a
homeserver.  Since to_device messages are delivered [exactly once] to each
device, this means that there's a race between A and B to receive and
acknowledge to_device messages. In the worst case, this can mean that one
client can "miss" to_device messages.

Instead, each client should register a unique (private) access token so that it
appears as a unique device to the homeserver. See also [client authentication]
in the spec.

[device]: https://matrix.org/docs/spec/index#devices
[exactly once]: https://matrix.org/docs/spec/client_server/r0.6.1#id70
[client authentication]: https://matrix.org/docs/spec/client_server/r0.6.1#client-authentication